Lazy initialization is a problem with singletons
Sometimes you only want the singleton to be initialized only when someone calls getInstance, and you don't want it
initialized in a static block or static context. Only made when someone needs it.

Potential race conditions if not protected against

Here we have two possible solutions to ensure thread safety:
1. synchronized getInstance method
2. Double-check locking - we use synchronized block on the class and check again.