- Examples to explore the Composite Design patter
The goal is to treat objects OR groups of objects in a uniform fashion.

It can behave as its own individual (scalar) Graphic object, or because we have a list of itself, and we have this
recursive nature, it can also act as a container or collection of Graphic objects.